refactor: 다중 노드 요약 모달 및 프로젝트 상세 페이지 훅 분리
