+++ frontend/src/components/projects/project-detail/multi-node-summary/MultiNodeSummaryModalContent.tsx
@@ -1,0 +1,99 @@
+'use client';
+
+import { IconClose, IconDownload } from '@wanteddev/wds-icon';
+import { useMemo, useRef } from 'react';
+
+import { buildMermaidCode, parseDevelopmentIdeas } from '@/utils/mermaid';
+
+import { ActionItemsSection } from './ActionItemsSection';
+import { DevelopmentIdeasSection } from './DevelopmentIdeasSection';
+import { MeetingRelationshipsSection } from './MeetingRelationshipsSection';
+import { ReferenceNodesSection } from './ReferenceNodesSection';
+import { RelationDiagramSection } from './RelationDiagramSection';
+import type { MultiNodeSummaryNode, MultiNodeSummaryResult } from './types';
+import { useSummaryPdfDownload } from './useSummaryPdfDownload';
+
+interface MultiNodeSummaryModalContentProps {
+  nodes: readonly MultiNodeSummaryNode[];
+  result: MultiNodeSummaryResult;
+  onDownloadClick?: () => void;
+  onClose: () => void;
+}
+
+export const MultiNodeSummaryModalContent = ({
+  nodes,
+  result,
+  onDownloadClick,
+  onClose,
+}: MultiNodeSummaryModalContentProps) => {
+  const pdfAreaRef = useRef<HTMLDivElement>(null);
+  const { isDownloading, handleDownloadClick } = useSummaryPdfDownload({
+    areaRef: pdfAreaRef,
+    onDownloaded: onDownloadClick,
+  });
+
+  const ideas = useMemo(
+    () => parseDevelopmentIdeas(result.development_ideas),
+    [result.development_ideas],
+  );
+  const diagramCode = useMemo(
+    () => buildMermaidCode(result.meeting_relationships),
+    [result.meeting_relationships],
+  );
+
+  return (
+    <>
+      {isDownloading && (
+        <div
+          className="bg-static-white/80 fixed inset-0 z-[10000] flex items-center justify-center backdrop-blur-[2px]"
+          role="status"
+          aria-live="polite"
+        >
+          <div className="flex flex-col items-center gap-3">
+            <span className="border-primary-40 inline-block h-8 w-8 animate-spin rounded-full border-2 border-t-transparent" />
+            <span className="text-body-1 text-label-normal font-medium">PDF를 준비하고 있어요</span>
+          </div>
+        </div>
+      )}
+      <div className="flex w-full flex-col gap-8">
+        {/* 헤더 */}
+        <div className="flex items-center justify-between">
+          <h2 className="text-heading-1 text-label-normal font-medium">AI 다중 노드 요약</h2>
+          <div className="flex items-center gap-3">
+            <button
+              type="button"
+              onClick={handleDownloadClick}
+              disabled={isDownloading}
+              aria-label="요약 PDF 다운로드"
+              className="text-label-alternative hover:text-label-neutral flex h-6 w-6 items-center justify-center border-none bg-transparent p-0 disabled:cursor-not-allowed disabled:opacity-50"
+            >
+              <IconDownload className="h-6 w-6" aria-hidden="true" />
+            </button>
+            <button
+              type="button"
+              onClick={onClose}
+              aria-label="닫기"
+              className="text-label-alternative hover:text-label-neutral flex h-6 w-6 items-center justify-center border-none bg-transparent p-0"
+            >
+              <IconClose className="h-6 w-6" aria-hidden="true" />
+            </button>
+          </div>
+        </div>
+
+        {/* 본문 — PDF 캡처 대상 영역 */}
+        <div
+          ref={pdfAreaRef}
+          className="custom-scrollbar flex max-h-150 w-full flex-col gap-6 overflow-y-auto pr-2"
+        >
+          <ReferenceNodesSection nodes={nodes} />
+          <ActionItemsSection analysis={result.action_items_analysis} />
+          <MeetingRelationshipsSection relationships={result.meeting_relationships} />
+          <DevelopmentIdeasSection ideas={ideas} />
+          <RelationDiagramSection code={diagramCode} />
+        </div>
+      </div>
+    </>
+  );
+};
+
+MultiNodeSummaryModalContent.displayName = 'MultiNodeSummaryModalContent';
